T: create test to check stub creation with `raw const` operator
Also, extract common code of the corresponding tests into `RsLazyBlockStubCreationTest`
